requirements.txt
# Arvis Room Intelligence System - Python Dependencies

# =============================================================================
# Core
# =============================================================================
python-dotenv>=1.0.0
loguru>=0.7.0

# =============================================================================
# Audio I/O
# =============================================================================
sounddevice>=0.4.6
numpy>=1.24.0
pygame>=2.5.0

# =============================================================================
# AI / Cloud APIs
# =============================================================================
openai>=1.0.0
httpx>=0.25.0

# =============================================================================
# Wake Word Detection
# =============================================================================
pvporcupine>=3.0.0

# =============================================================================
# Vision / Camera (Phase 3)
# =============================================================================
opencv-python>=4.8.0
mediapipe>=0.10.0

# =============================================================================
# Scheduling
# =============================================================================
schedule>=1.2.0

# =============================================================================
# GPIO / Hardware (Raspberry Pi - uncomment when on Pi)
# =============================================================================
# RPi.GPIO>=0.7.1
# gpiozero>=2.0
# rpi_ws281x>=5.0.0

# =============================================================================
# Smart Plug Control
# =============================================================================
python-kasa>=0.7.7

# =============================================================================
# Spotify (Phase 5 - optional)
# =============================================================================
# spotipy>=2.23.0

# =============================================================================
# Testing
# =============================================================================
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
